pyproject.toml
[project]
name = "evm-sleuth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dlt>=1.14.1",
    "eth-hash[pycryptodome]>=0.7.1",
    "jupyter>=1.1.1",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]

[tool.uv.sources]
evm_sleuth = { path = "./src/evm_sleuth" }

[tool.uv]
package = true
